GDALRasterSource is a case class now
